Address comments from clean up PR
