+++ envoyfilter/src/common_types.rs
@@ -52,3 +52,20 @@ pub struct VectorPoint {
 pub struct CreateVectorStorePoints {
     pub points: Vec<VectorPoint>,
 }
+
+pub mod open_ai {
+    use serde::{Deserialize, Serialize};
+
+    #[derive(Debug, Clone, Serialize, Deserialize)]
+    pub struct ChatCompletions {
+        #[serde(default)]
+        pub model: String,
+        pub messages: Vec<Message>,
+    }
+
+    #[derive(Debug, Clone, Serialize, Deserialize)]
+    pub struct Message {
+        pub role: String,
+        pub content: String,
+    }
+}
